+++ rex/modules/affine.py
@@ -1,0 +1,41 @@
+import torch
+import torch.nn as nn
+
+
+class Biaffine(nn.Module):
+    """Biaffine transformation
+
+    References:
+        - https://github.com/yzhangcs/parser/blob/main/supar/modules/affine.py
+    """
+
+    def __init__(self, n_in, n_out=2, bias_x=True, bias_y=True):
+        super().__init__()
+
+        self.n_in = n_in
+        self.n_out = n_out
+        self.bias_x = bias_x
+        self.bias_y = bias_y
+        weight = torch.zeros(n_out, n_in + int(bias_x), n_in + int(bias_y))
+        nn.init.xavier_normal_(weight)
+        self.weight = nn.Parameter(weight, requires_grad=True)
+
+    def extra_repr(self):
+        s = f"n_in={self.n_in}, n_out={self.n_out}"
+        if self.bias_x:
+            s += f", bias_x={self.bias_x}"
+        if self.bias_y:
+            s += f", bias_y={self.bias_y}"
+
+        return s
+
+    def forward(self, x, y):
+        if self.bias_x:
+            x = torch.cat((x, torch.ones_like(x[..., :1])), -1)
+        if self.bias_y:
+            y = torch.cat((y, torch.ones_like(y[..., :1])), -1)
+        # [batch_size, n_out, seq_len, seq_len]
+        s = torch.einsum("bxi,oij,byj->boxy", x, self.weight, y)
+        # s = s.permute(0, 2, 3, 1)
+
+        return s